+++ src/helpers.rs
@@ -176,17 +176,34 @@ pub fn check_file_sha1_hash(
 #[cfg(test)]
 mod tests {
     use super::*;
+    use std::io::Write;
+    use tempfile::NamedTempFile;
 
     #[test]
-    // TODO
     fn check_file_sha1_hash_ok() {
-        assert_eq!(0, 1);
+        // Create test file of known sha1sum hash
+        let mut tmpfile = NamedTempFile::new().expect("Could not create tmp file");
+        write!(tmpfile, "Hello World!").unwrap();
+        tmpfile.flush().expect("Error flushing tmp file");
+
+        // Check the checksum of the tmp file
+        assert!(check_file_sha1_hash(
+            &tmpfile.path().to_path_buf(),
+            "2ef7bde608ce5404e97d5f042f95f89f1c232871",
+            123
+        )
+        .is_ok());
     }
 
     #[test]
-    // TODO
     fn check_file_sha1_hash_err() {
-        assert_eq!(0, 1);
+        // Create test file of known sha1sum hash
+        let mut tmpfile = NamedTempFile::new().expect("Could not create tmp file");
+        write!(tmpfile, "Hello World!").unwrap();
+        tmpfile.flush().expect("Error flushing tmp file");
+
+        // Check the checksum of the tmp file - but the expected checksum is wrong
+        assert!(check_file_sha1_hash(&tmpfile.path().to_path_buf(), "abcd123", 123).is_err());
     }
 
     #[test]